\name{zelig2arima}
\alias{zelig2arima}
\title{Interface between the Zelig Model arima and
the Pre-existing Model-fitting Method}
\usage{
  zelig2arima(formula, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model arima and the
  Pre-existing Model-fitting Method
}

