\name{zelig2cox.ph}
\alias{zelig2cox.ph}
\title{Interface between the Zelig Model coxph and...}
\usage{zelig2cox.ph(formula, ..., robust=FALSE, cluster, data)}
\description{Interface between the Zelig Model coxph and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
