\name{zelig2cox.ph}
\alias{zelig2cox.ph}
\title{Interface between the Zelig Model coxph and
the Pre-existing Model-fitting Method}
\usage{
  zelig2cox.ph(formula, ..., robust = FALSE, cluster, data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model coxph and the
  Pre-existing Model-fitting Method
}

