\name{zelig2ei.RxC}
\alias{zelig2ei.RxC}
\title{Interface between the Zelig Model ei...}
\usage{zelig2ei.RxC(formula, ..., data)}
\description{Interface between the Zelig Model ei.RxC and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
