\name{zelig2ei.RxC}
\alias{zelig2ei.RxC}
\title{Interface between the Zelig Model ei.RxC and
the Pre-existing Model-fitting Method}
\usage{
  zelig2ei.RxC(formula, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model ei.RxC and the
  Pre-existing Model-fitting Method
}

