\docType{package}
\name{ZeligMisc-package}
\alias{ZeligMisc}
\alias{ZeligMisc-package}
\title{ZeligMisc}
\description{
  \tabular{ll}{ Package: \tab ZeligMisc\cr Version: \tab
  0.1\cr Date: \tab 2011-04-25\cr Depends: Zelig License:
  \tab GPL version 2 or newer\cr }
}
\details{
  Edit this description
}
\author{
  Matt Owen
}
\keyword{package}

