\name{zelig2arima}
\alias{zelig2arima}
\title{Interface between the Zelig Model arima and...}
\usage{zelig2arima(formula, ..., data)}
\description{Interface between the Zelig Model arima and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
