\name{ZeligMisc-package}
\title{ZeligMisc...}
\description{ZeligMisc}
\details{\tabular{ll}{
Package: \tab ZeligMisc\cr
Version: \tab 0.1\cr
Date: \tab 2011-04-25\cr
Depends: Zelig
License: \tab GPL version 2 or newer\cr
}

Edit this description}
\alias{ZeligMisc-package}
\alias{ZeligMisc}
\docType{package}
\author{Matt Owen}
\keyword{package}
